+++ allennlp/modules/token_embedders/pretrained_transformer_embedder.py
@@ -38,7 +38,9 @@ class PretrainedTransformerEmbedder(TokenEmbedder):
         as embedders such as BERT. However, other models consist of encoder and decoder, in which case we just
         want to use the encoder.
     train_parameters: `bool`, optional (default = `True`)
-        If this is `True`, the transformer weights get updated during training.
+        If this is `True`, the transformer weights get updated during training. If this is `False`, the
+        transformer weights are not updated during training and its dropout and batch normalization layers
+        are set to evaluation mode.
     last_layer_only: `bool`, optional (default = `True`)
         When `True` (the default), only the final layer of the pretrained transformer is taken
         for the embeddings. But if set to `False`, a scalar mix of all of the layers
@@ -122,6 +124,7 @@ class PretrainedTransformerEmbedder(TokenEmbedder):
         self._num_added_tokens = self._num_added_start_tokens + self._num_added_end_tokens
 
         if not train_parameters:
+            self.transformer_model.eval()
             for param in self.transformer_model.parameters():
                 param.requires_grad = False